queue v0.2.0 Queue

Очередь сообщений. Перед началом использования необходимо остановить текущий сервис. Выполнить инициализацию Mnesia с помощью команды mix queue.setup

Functions

ack(ids)

ack([integer()] | integer()) :: :ok

Сообщить об успешном выполнение. Элементы будут удалены из очереди.

Параметры

ids - список id элементов, которые завершились успехом. ID можно передавать и по одному.

Возвращает

Атом :ok в случае успеха, {:error, reason} в противном случае.

Пример

iex> Queue.add([1,2])
iex> [%{id: id1, value: value1} | _T] = Queue.get()
iex> value1
1
iex> [%{id: id2, value: value2} | _T] = Queue.get()
iex> value2
2
iex> :ok = Queue.ack(id2)
iex> Queue.get()
:empty

iex> Queue.add([1,2,3])
iex> Queue.get(3) |> Enum.map(fn %{id: id} -> id end) |> Queue.reject
iex> [%{id: id, value: value} | _T] = Queue.get()
iex> value
1

add(vals)

add(any()) :: :ok

Добавить элемент в конец очереди.

Параметры

vals - значение/значения, добавляемые в конец очереди.

Пример

iex> Queue.add(1)
:ok

iex> Queue.add([1,2,3,4,5,6])
:ok

get(limit \\ 1)

get(integer()) :: {:error, any()} | Queue.DB.Jobs.result()

Получить элемент/элементы из очереди.

Параметры

limit - количество получаемых элементов, по умолчанию 1.

Возвращает

Лист с элементами, каждый элемент представлен ассоциативном массивом (map) с ключами id и value.

Пример

iex> Queue.get()
:empty

iex> :ok = Queue.add(123)
iex> [%{id: id, value: value} | _T] = Queue.get()
iex> value
123

reject(ids)

reject([integer()] | integer()) :: :ok

Сообщить о неудачном выполнение. Элементы будут возвращены в конец очереди.

Параметры

ids - список id элементов, которые завершились не удачей. ID можно передавать и по одному.

Возвращает

Атом :ok в случае успеха, {:error, reason} в противном случае.

Пример

iex> Queue.add([1,2,3])
iex> [%{id: id1, value: value1} | _T] = Queue.get()
iex> value1
1
iex> [%{id: id2, value: value2} | _T] = Queue.get()
iex> value2
2
iex> Queue.reject(id2)
iex> [%{id: id3, value: value3} | _T] = Queue.get()
iex> value3
3
iex> [%{id: id4, value: value4} | _T] = Queue.get()
iex> value4
2

iex> Queue.add([1,2,3])
iex> Queue.get(3) |> Enum.map(fn %{id: id} -> id end) |> Queue.reject
iex> [%{id: id, value: value} | _T] = Queue.get()
iex> value
1

start(type, args)

start(any(), any()) :: :ignore | {:error, any()} | {:ok, pid()}

Called when an application is started.

This function is called when an application is started using Application.start/2 (and functions on top of that, such as Application.ensure_started/2). This function should start the top-level process of the application (which should be the top supervisor of the application's supervision tree if the application follows the OTP design principles around supervision).

start_type defines how the application is started:

:normal - used if the startup is a normal startup or if the application is distributed and is started on the current node because of a failover from another node and the application specification key :start_phases is :undefined.

{:takeover, node} - used if the application is distributed and is started on the current node because of a failover on the node node.

{:failover, node} - used if the application is distributed and is started on the current node because of a failover on node node, and the application specification key :start_phases is not :undefined.

start_args are the arguments passed to the application in the :mod specification key (e.g., mod: {MyApp, [:my_args]}).

This function should either return {:ok, pid} or {:ok, pid, state} if startup is successful. pid should be the PID of the top supervisor. state can be an arbitrary term, and if omitted will default to []; if the application is later stopped, state is passed to the stop/1 callback (see the documentation for the c:stop/1 callback for more information).

use Application provides no default implementation for the start/2 callback.

Callback implementation for Application.start/2.