queue v0.1.0 Queue

Персистентная очередь сообщений. Очередь реализована с использованием двух стеков.

Пример

iex> Queue.add(1)
iex> Queue.add(2)
iex> {val, _} = Queue.get()
iex> val
1
iex> Queue.reject()
iex> {val, _} = Queue.get()
iex> val
2
iex> {val, _} = Queue.get()
iex> val
1

Types

stack()

stack() :: {:stack, reference(), any(), reference()}

t()

t() :: {:queue, reference(), reference(), reference()}

Functions

ack()

ack() :: :ok

Подтвердить успешную обработку сообщения. После подтверждения, элемент пропадет из временого буфера. Откатит такой элемент уже будет не возможно.

Пример

iex> Queue.add(1)
iex> Queue.get()
iex> Queue.ack()
:ok

add(val)

add(any()) :: reference()

Добавить элемент в конец очереди.

Параметры

val - значение, добавленное в конец очереди.

Возвращает

Ссылку на новую очередь.

Пример

iex> new_ref = Queue.add(1)
iex> is_reference(new_ref)
true

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get()

get() :: {any(), reference()}

Получить элемент из очереди.

Возвращает

Элемент и ссылку {val, new_ref} на новую очередь. В случае если очередь пустая, будет возвращено {:empty, new_ref}.

Элемент помещается во временный буфер и закрепляется за процессом, получившим его. В дальнейшим процесс может подтвердить выполнение работы: ack/0 или сообщить об ошибке обработки reject/0.

Пример

iex> {:empty, _} = Queue.get()
iex> _ = Queue.add(1)
iex> {val, _} = Queue.get()
iex> val
1

init(opts)

init(nil) :: {:ok, {reference(), map()}}

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

init_arg is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state} except handle_info(:timeout, state) will be called after timeout milliseconds if no messages are received within the timeout.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

queue(args \\ []) (macro)

Структура очереди на основе двух стеков. Представлена в виде записи, с полями:

ref - уникальная сссылка текущей очереди.

l - ссылка на левый стек.

r - ссылка на правый стек.

queue(record, args) (macro)

reject()

reject() :: :ok

Откатить все полученные элементы. Элементы будут возвращены в конец очереди.

Пример

iex> Queue.add(1)
iex> {val, _} = Queue.get()
iex> val
1
iex> Queue.add(2)
iex> _ = Queue.reject()
iex> {val, _} = Queue.get
iex> val
2
iex> {val, _} = Queue.get
iex> val
1

rollback(ref)

rollback(reference()) :: reference()

Откатит очередь к некоторму состоянию.

Пример

iex> q1 = Queue.add(1)
iex> {val, _} = Queue.get
iex> val
1
iex> Queue.rollback(q1)
iex> {val, _} = Queue.get
iex> val
1

stack(args \\ []) (macro)

Структура стека. Представлена в виде записи, с полями:

ref - уникальная ссылка текущего стека.

val - текущее значение.

pref - ссылка на предыдущее значение стека.

stack(record, args) (macro)

start_link()

start_link() :: :ignore | {:error, any()} | {:ok, pid()}